Time: 13 ms (78.54%), Space: 12.2 MB (21.73%) - LeetHub

+++ 150-evaluate-reverse-polish-notation/150-evaluate-reverse-polish-notation.cpp
@@ -1,48 +1,48 @@
 class Solution {
 public:
     int evalRPN(vector<string>& tokens) {
-         stack<int> st;
-        for(int i=0;i<tokens.size();i++)
+         stack<long long int> st;
+        for(long long int i=0;i<tokens.size();i++)
         {
             if(tokens[i]!="*" && tokens[i]!="-" && tokens[i]!="+" && tokens[i]!="/")
             {
-                int num=stoi(tokens[i]);
+                long long int num=stoi(tokens[i]);
                 st.push(num);
             }
             else
             {
                 if(tokens[i]=="+")
                 {
-                    int b=st.top();
+                    long long int b=st.top();
                     st.pop();
-                    int c=st.top();
+                    long long int c=st.top();
                     st.pop();
                     c=b+c;
                     st.push(c);
                 }
                 else if(tokens[i]=="-")
                 {
-                    int b=st.top();
+                    long long int b=st.top();
                     st.pop();
-                    int c=st.top();
+                    long long int c=st.top();
                     st.pop();
                     c=c-b;
                     st.push(c);
                 }
                 else if(tokens[i]=="*")
                 {
-                    int b=st.top();
+                    long long int b=st.top();
                     st.pop();
-                    int c=st.top();
+                    long long int c=st.top();
                     st.pop();
                     c=c*b;
                     st.push(c);
                 }
                 else
                 {
-                    int b=st.top();
+                    long long int b=st.top();
                     st.pop();
-                    int c=st.top();
+                    long long int c=st.top();
                     st.pop();
                     c=c/b;
                     st.push(c);
